image-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
pytesseract==0.3.10
torch==2.1.2
torchvision==0.16.2
git+https://github.com/openai/CLIP.git
ftfy==6.1.3
regex==2023.12.25